break circular includes between transform.h/boundbox.h/volume_stack.h; const-ref GetWorldBBox
